requirements.txt
diffusers==0.18.2
catalyst
docker==7.1.0
gym==0.26.2
huggingface-hub==0.25.0
imageio==2.36.0
ipython==8.18.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.7
matplotlib==3.9.2
mlflow==2.17.1
mlflow-skinny==2.17.1
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pygame==2.1.2
python-box==7.2.0
PyYAML==6.0.2
scikit-image==0.19.3
scikit-learn==1.5.2
scikit-video==1.1.11
scipy==1.13.1
Shapely==1.8.4
six==1.16.0
tensorboardX==2.6.2.2
torch==1.13.1
torchdiffeq==0.2.4
torchvision==0.14.1
tqdm==4.66.5
torch-kmeans
pynvml
onnxruntime
onnxruntime-gpu
